requirements.txt
# =============================================================================
# SOC Copilot — Python Dependencies
# Generated from pyproject.toml
# Install: pip install -r requirements.txt
# =============================================================================

# --- Core ML ---
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
joblib>=1.3.0

# --- Configuration & Validation ---
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0.0

# --- Log Parsing ---
python-evtx>=0.8.0

# --- Desktop UI ---
PyQt6>=6.5.0

# --- External Security APIs ---
requests>=2.28.0

# --- Utilities ---
structlog>=23.1.0
python-dateutil>=2.8.0
